+++ govcd/edgegateway_test.go
@@ -7,6 +7,10 @@
 package govcd
 
 import (
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"regexp"
 	"strings"
 
 	"github.com/vmware/go-vcloud-director/v2/types/v56"
@@ -605,18 +609,18 @@ func (vcd *TestVCD) TestEdgeGateway_UpdateFwGeneralParams(check *C) {
 	// Cache current firewall settings for change validation in the end
 	beforeFw, beforeFwXml := testCacheFirewall(*edge, check)
 
-	_, err = edge.UpdateFwGeneralParams(false, false, "deny")
+	_, err = edge.UpdateFwGeneralConfig(false, false, "deny")
 	check.Assert(err, IsNil)
 
-	_, err = edge.UpdateFwGeneralParams(true, true, "accept")
+	_, err = edge.UpdateFwGeneralConfig(true, true, "accept")
 	check.Assert(err, IsNil)
 
 	// Try to set invalid loglevel to get validation error
-	_, err = edge.UpdateFwGeneralParams(false, false, "invalid_action")
+	_, err = edge.UpdateFwGeneralConfig(false, false, "invalid_action")
 	check.Assert(err, ErrorMatches, ".*default action must be either 'accept' or 'deny'.*")
 
 	// Restore to initial settings and validate that it
-	_, err = edge.UpdateFwGeneralParams(beforeFw.Enabled, beforeFw.DefaultPolicy.LoggingEnabled, beforeFw.DefaultPolicy.Action)
+	_, err = edge.UpdateFwGeneralConfig(beforeFw.Enabled, beforeFw.DefaultPolicy.LoggingEnabled, beforeFw.DefaultPolicy.Action)
 	check.Assert(err, IsNil)
 
 	// Validate configuration against initially cached version
@@ -658,4 +662,112 @@ func (vcd *TestVCD) TestEdgeGateway_GetVnics(check *C) {
 	}
 	check.Assert(foundExtNet, Equals, true)
 	check.Assert(foundOrgNet, Equals, true)
+}
+
+// testGetEdgeEndpointXML is used for additional validation that modifying edge gateway endpoint
+// does not change any single field. It returns an XML string of whole configuration
+func testGetEdgeEndpointXML(endpoint string, edge EdgeGateway, check *C) string {
+
+	httpPath, err := edge.buildProxiedEdgeEndpointURL(endpoint)
+	check.Assert(err, IsNil)
+
+	resp, err := edge.client.ExecuteRequestWithCustomError(httpPath, http.MethodGet, types.AnyXMLMime,
+		fmt.Sprintf("unable to get XML from endpoint %s: %%s", endpoint), nil, &types.NSXError{})
+	check.Assert(err, IsNil)
+
+	defer resp.Body.Close()
+
+	body, err := ioutil.ReadAll(resp.Body)
+	check.Assert(err, IsNil)
+
+	return string(body)
+}
+
+// cacheLoadBalancer is meant to store load balancer settings before any operations so that all
+// configuration can be checked after manipulation
+func testCacheLoadBalancer(edge EdgeGateway, check *C) (*types.LbGeneralParamsWithXml, string) {
+	beforeLb, err := edge.GetLBGeneralParams()
+	check.Assert(err, IsNil)
+	beforeLbXml := testGetEdgeEndpointXML(types.LbConfigPath, edge, check)
+	return beforeLb, beforeLbXml
+}
+
+// testCacheFirewall is meant to store firewall settings before any operations so that all
+// configuration can be checked after manipulation
+func testCacheFirewall(edge EdgeGateway, check *C) (*types.FwGeneralParamsWithXml, string) {
+	beforeFw, err := edge.GetFwGeneralConfig()
+	check.Assert(err, IsNil)
+	beforeFwbXml := testGetEdgeEndpointXML(types.EdgeFirewallPath, edge, check)
+	return beforeFw, beforeFwbXml
+}
+
+// testCheckLoadBalancerConfig validates if both raw XML string and load balancer struct remain
+// identical after settings manipulation.
+func testCheckLoadBalancerConfig(beforeLb *types.LbGeneralParamsWithXml, beforeLbXml string, edge EdgeGateway, check *C) {
+	afterLb, err := edge.GetLBGeneralParams()
+	check.Assert(err, IsNil)
+
+	afterLbXml := testGetEdgeEndpointXML(types.LbConfigPath, edge, check)
+
+	// remove `<version></version>` tag from both XML represntation and struct for deep comparison
+	// because this version changes with each update and will never be the same after a few
+	// operations
+
+	reVersion := regexp.MustCompile(`<version>\w*<\/version>`)
+	beforeLbXml = reVersion.ReplaceAllLiteralString(beforeLbXml, "")
+	afterLbXml = reVersion.ReplaceAllLiteralString(afterLbXml, "")
+
+	beforeLb.Version = ""
+	afterLb.Version = ""
+
+	check.Assert(beforeLb, DeepEquals, afterLb)
+	check.Assert(beforeLbXml, DeepEquals, afterLbXml)
+}
+
+// testCheckFirewallConfig validates if both raw XML string and firewall struct remain
+// identical after settings manipulation.
+func testCheckFirewallConfig(beforeFw *types.FwGeneralParamsWithXml, beforeFwXml string, edge EdgeGateway, check *C) {
+	afterFw, err := edge.GetFwGeneralConfig()
+	check.Assert(err, IsNil)
+
+	afterFwXml := testGetEdgeEndpointXML(types.EdgeFirewallPath, edge, check)
+
+	// remove `<version></version>` tag from both XML represntation and struct for deep comparison
+	// because this version changes with each update and will never be the same after a few
+	// operations
+	reVersion := regexp.MustCompile(`<version>\w*<\/version>`)
+	beforeFwXml = reVersion.ReplaceAllLiteralString(beforeFwXml, "")
+	afterFwXml = reVersion.ReplaceAllLiteralString(afterFwXml, "")
+
+	beforeFw.Version = ""
+	afterFw.Version = ""
+
+	// Because the test enables and disables firewall configuration, main firewall rule has its ID
+	// and ruleTag changed during the test and we must ignore this change while comparing. This is
+	// always the first rule kept at the top so we are replacing `id` and `ruletag` fields in
+	// "before" and "after" values.
+	beforeFw.FirewallRules.Text = replaceFirstMatch(beforeFw.FirewallRules.Text, `<id>\d*</id>`, "<id>99999</id>")
+	beforeFw.FirewallRules.Text = replaceFirstMatch(beforeFw.FirewallRules.Text, `<ruleTag>\d*</ruleTag>`, "<ruleTag>99999</ruleTag>")
+	afterFw.FirewallRules.Text = replaceFirstMatch(afterFw.FirewallRules.Text, `<id>\d*</id>`, "<id>99999</id>")
+	afterFw.FirewallRules.Text = replaceFirstMatch(afterFw.FirewallRules.Text, `<ruleTag>\d*</ruleTag>`, "<ruleTag>99999</ruleTag>")
+
+	beforeFwXml = replaceFirstMatch(beforeFwXml, `<id>\d*</id>`, "<id>99999</id>")
+	beforeFwXml = replaceFirstMatch(beforeFwXml, `<ruleTag>\d*</ruleTag>`, "<ruleTag>99999</ruleTag>")
+	afterFwXml = replaceFirstMatch(afterFwXml, `<id>\d*</id>`, "<id>99999</id>")
+	afterFwXml = replaceFirstMatch(afterFwXml, `<ruleTag>\d*</ruleTag>`, "<ruleTag>99999</ruleTag>")
+
+	check.Assert(beforeFw, DeepEquals, afterFw)
+	check.Assert(beforeFwXml, DeepEquals, afterFwXml)
+}
+
+// replaceFirstMatch replaces first `regex` matched in `text` with `replacement` and returns it
+// It will panic if regex is invalid.
+func replaceFirstMatch(text, regex, replacement string) string {
+	re := regexp.MustCompile(regex)
+	// Replace leftmost found string
+	found := re.FindString(text)
+	if found != "" {
+		return strings.Replace(text, found, replacement, 1)
+	}
+	return ""
 }
